+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project are documented in this file.
 
-## [1.1.5] - unreleased
+## [1.1.5] - 2026-08-15
 
 ### Added